Use strong type in code for performance

+++ R/NestedCategBayesImpute/src/checkconstraints.h
@@ -16,5 +16,5 @@
 #define AGE 3
 #define RELATE 4
 
-int checkconstraints_imp(double *data, double *isPossible,int hh_size, int DIM, int nHouseholds);
-int checkconstraints_imp_HHhead_at_group_level(double *data, double *isPossible,int hh_size, int DIM, int nHouseholds);
+int checkconstraints_imp(int *data, int *isPossible,int hh_size, int DIM, int nHouseholds);
+int checkconstraints_imp_HHhead_at_group_level(int *data, int *isPossible,int hh_size, int DIM, int nHouseholds);
